# documentation/API/study-datastore/swagger.json
{
  "swagger" : "2.0",
  "info" : {
    "version" : "2.0.3",
    "title" : "Study Meta Data",
    "termsOfService" : "http://www.github.com/kongchen/swagger-maven-plugin",
    "license" : {
      "name" : "Apache 2.0",
      "url" : "http://www.apache.org/licenses/LICENSE-2.0.html"
    }
  },
...
  "schemes" : [ "http", "https" ],
  "paths" : {
...
    "/consentDocument" : {
      "get" : {
        "tags" : [ "Studies" ],
        "summary" : "Get the consent Document for a particular study based on the consent Version",
        "description" : "",
        "operationId" : "consentDocument",
        "consumes" : [ "application/json" ],
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "name" : "studyId",
          "in" : "query",
          "required" : true,
          "type" : "string"
        }, {
          "name" : "consentVersion",
          "in" : "query",
          "required" : true,
          "type" : "string"
        }, {
          "name" : "activityId",
          "in" : "query",
          "required" : true,
          "type" : "string"
        }, {
          "name" : "activityVersion",
          "in" : "query",
          "required" : true,
          "type" : "string"
        } ],
        "responses" : {
          "102" : {
            "description" : "Invalid input."
          },
          "103" : {
            "description" : "No records found"
          },
          "104" : {
            "description" : "UNKNOWN"
          },
          "200" : {
            "description" : "Successful operation",
            "schema" : {
              "$ref" : "#/definitions/ConsentDocumentResponse"
            }
          }
        }
      }
    },
...
    "/studyInfo" : {
      "get" : {
        "tags" : [ "Studies" ],
        "summary" : "Get the study information for a particular study",
        "description" : "",
        "operationId" : "studyInfo",
        "consumes" : [ "application/json" ],
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "name" : "studyId",
          "in" : "query",
          "required" : true,
          "type" : "string"
        } ],
        "responses" : {
          "102" : {
            "description" : "Invalid input."
          },
          "103" : {
            "description" : "No records found"
          },
          "104" : {
            "description" : "UNKNOWN"
          },
          "200" : {
            "description" : "Successful operation",
            "schema" : {
              "$ref" : "#/definitions/StudyInfoResponse"
            }
          }
        }
      }
    },
...
  },
  "definitions" : {
...
    "AnchorDateBean" : {
      "type" : "object",
      "properties" : {
        "type" : {
          "type" : "string"
        },
        "questionInfo" : {
          "$ref" : "#/definitions/QuestionInfoBean"
        }
      }
    },
...
    "ConsentDocumentBean" : {
      "type" : "object",
      "properties" : {
        "version" : {
          "type" : "string"
        },
        "type" : {
          "type" : "string"
        },
        "content" : {
          "type" : "string"
        }
      }
    },
    "ConsentDocumentResponse" : {
      "type" : "object",
      "properties" : {
        "message" : {
          "type" : "string"
        },
        "consent" : {
          "$ref" : "#/definitions/ConsentDocumentBean"
        }
      }
    },
...
    "InfoBean" : {
      "type" : "object",
      "properties" : {
        "type" : {
          "type" : "string"
        },
        "image" : {
          "type" : "string"
        },
        "title" : {
          "type" : "string"
        },
        "text" : {
          "type" : "string"
        },
        "videoLink" : {
          "type" : "string"
        }
      }
    },
...
    "QuestionInfoBean" : {
      "type" : "object",
      "properties" : {
        "activityId" : {
          "type" : "string"
        },
        "activityVersion" : {
          "type" : "string"
        },
        "key" : {
          "type" : "string"
        }
      }
    },
...
    "StudyInfoResponse" : {
      "type" : "object",
      "properties" : {
        "message" : {
          "type" : "string"
        },
        "studyWebsite" : {
          "type" : "string"
        },
        "info" : {
          "type" : "array",
          "items" : {
            "$ref" : "#/definitions/InfoBean"
          }
        },
        "anchorDate" : {
          "$ref" : "#/definitions/AnchorDateBean"
        }
      }
    },
...
  }
}